+++ src/components/headers/Header2.jsx
@@ -8,11 +8,13 @@ import { FiFilter } from 'react-icons/fi';
 import BarraDePesquisa from './BarraDePesquisa';
 import { useAuth } from '../../features/auth/useAuth';
 import ModalAvaliacao from '../ModalAvaliacao';
+import ModalFiltroAvaliacao from '../ModalFiltroAvaliacoes';
 
 //Padrão de Header com logo com título, barra de pesquisa expansível, botão de filtrar e botão de sair
-function Header({ placeholder }) {
+function Header({ placeholder, onAplicarFiltros, filtros, ordenador }) {
   const [mostrarPesquisa, setMostrarPesquisa] = useState(false);
   const [modalVisivel, setModalVisivel] = useState(false);
+  const [modalFiltroVisivel, setModalFiltroVisivel] = useState(false);
   const { logout } = useAuth();
   const { user } = useAuth();
   const location = useLocation();
@@ -71,10 +73,20 @@ function Header({ placeholder }) {
                 <IoSearch className="w-[30px] sm:w-[30px] md:w-[40px] lg:w-[45px] h-auto" alt="Pesquisar" />
               )}
             </button>
-            <button className="flex-shrink-0">
+            <button className="flex-shrink-0" onClick={() => setModalFiltroVisivel(true)}>
               <FiFilter className="w-[30px] sm:w-[30px] md:w-[40px] lg:w-[45px] h-auto" alt="Filtrar" />
             </button>
           </>
+        )}
+
+        {modalFiltroVisivel && location.pathname == '/feed' && (
+          <ModalFiltroAvaliacao
+            open={modalFiltroVisivel}
+            onClose={() => setModalFiltroVisivel(false)}
+            onAplicar={onAplicarFiltros}
+            filtros={filtros}
+            ordenador={ordenador}
+          />
         )}
 
         {mostrarBotaoNovaAvaliacao && (